Split NomenclatureCategory into separate file
